# mobilapp/docs/api-integration/docker/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FitnessApp API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/me/goals/current": {
      "get": {
        "tags": [
          "Goals"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/me/goals/at": {
      "get": {
        "tags": [
          "Goals"
        ],
        "parameters": [
          {
            "name": "at",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/me/goals/{goalId}": {
      "get": {
        "tags": [
          "Goals"
        ],
        "parameters": [
          {
            "name": "goalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/me/goals": {
      "get": {
        "tags": [
          "Goals"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDtoCursorPage"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDtoCursorPage"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDtoCursorPage"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Goals"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserGoalRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserGoalRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserGoalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/me/goals/recalculate": {
      "post": {
        "tags": [
          "Goals"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoalDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserGoalRequest": {
        "type": "object",
        "properties": {
          "goalType": {
            "$ref": "#/components/schemas/GoalType"
          },
          "targetWeight": {
            "type": "number",
            "format": "double"
          },
          "weightChangePerWeek": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "GoalType": {
        "enum": [
          "LoseWeight",
          "MaintainWeight",
          "GainWeight"
        ],
        "type": "string"
      },
...
      "UserGoalDto": {
        "type": "object",
        "properties": {
          "userGoalId": {
            "type": "integer",
            "format": "int32"
          },
          "goalType": {
            "$ref": "#/components/schemas/GoalType"
          },
          "targetWeight": {
            "type": "number",
            "format": "double"
          },
          "weightChangePerWeek": {
            "type": "number",
            "format": "double"
          },
          "targetDailyCalories": {
            "type": "number",
            "format": "double"
          },
          "targetProtein": {
            "type": "number",
            "format": "double"
          },
          "targetCarbohydrates": {
            "type": "number",
            "format": "double"
          },
          "targetFat": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UserGoalDtoCursorPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserGoalDto"
            },
            "nullable": true
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter the access JWT.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}